+++ lib/fl_country_code_picker.dart
@@ -1,7 +1,8 @@
-/// Library for fl_country_code_picker package.
+/// Library for fl_country_code_picker.
 library fl_country_code_picker;
 
 export 'src/fl_country_code_picker.dart';
 export 'src/models/models.dart';
+export 'src/view/views.dart';
 export 'src/widgets/widgets.dart'
-    hide CcpDefaultListItemTrailing, CcpDefaultModalTitle, CcpDefaultSearchBar;
+    hide CcpDefaultListItemTrailing, CcpDefaultModalTitle, SearchBar;