+++ lib/src/widgets/widgets.dart
@@ -1,4 +1,3 @@
 export 'ccp_default_list_item_trailing.w.dart';
 export 'ccp_default_modal_title.w.dart';
-export 'ccp_default_search_bar.w.dart';
-export 'country_code_picker_modal.dart';
+export 'search_bar.w.dart';